Refactor separate CLI input and ouput
I'm not sure yet if this is a good idea, but I want
to extract input validation next and feel like the CLI
module file would become difficult to read if too long.
